Добавит мини фикацию скриптов и стилей
