+++ .babelrc
@@ -1,0 +1,4 @@
+// .babelrc
+{
+  "presets": ["@babel/preset-env"]
+}
