+++ js/script.js
@@ -1,199 +1,1 @@
-class Field {
-  constructor(element) {
-    if (!new.target) {
-      throw new Error("Field must be called with a new operator");
-    }
-    if (!element) {
-      throw new Error("new Field: element is empty");
-    }
-    this._element = element;
-    this._validators = [element => element.validity.valid];
-
-    this._element.addEventListener("keyup", () => {
-      this.valid();
-    });
-
-    this._element.addEventListener("change", () => {
-      this.valid();
-    });
-  }
-
-  setRequired() {
-    this._validators.unshift(() => this._element.value);
-  }
-
-  get value() {
-    return this._element.value;
-  }
-
-  set value(value) {
-    this._element.value = value;
-  }
-
-  valid() {
-    const isValid = this._validators.every(v => v(this._element));
-
-    this._element.classList.remove("input__field--invalid");
-    if (!isValid) {
-      this._element.classList.add("input__field--invalid");
-    }
-    return isValid;
-  }
-
-  focus() {
-    return this._element.focus();
-  }
-
-  addChangeHandler(handler) {
-    this._element.addEventListener("change", handler);
-  }
-}
-
-class Form {
-  constructor(element) {
-    if (!new.target) {
-      throw new Error("Form must be called with a new operator");
-    }
-    if (!element) {
-      throw new Error("new Form: element is empty");
-    }
-    if (!(element instanceof HTMLFormElement)) {
-      throw new Error("new Form: element is not a form element");
-    }
-    element.noValidate = true;
-
-    this._element = element;
-    this._fields = [];
-  }
-
-  addFields(...fields) {
-    this._fields.push(...fields);
-  }
-
-  valid() {
-    let valid = true;
-    this._fields.forEach(f => (valid &= f.valid()));
-
-    return valid;
-  }
-
-  get fields() {
-    return this._fields;
-  }
-
-  addSubmitHandler(handler) {
-    this._element.addEventListener("submit", handler);
-  }
-}
-
-class Modal {
-  constructor(element) {
-    if (!new.target) {
-      throw new Error("Modal must be called with a new operator");
-    }
-    this._element = element;
-
-    this._initClose();
-  }
-
-  _initClose() {
-    const close = this.close.bind(this);
-
-    window.addEventListener("keydown", evt => {
-      if (evt.keyCode !== 27) {
-        return;
-      }
-      evt.preventDefault();
-
-      close();
-    });
-
-    const closeButton = this._element.querySelector(".popup__close");
-    if (closeButton) {
-      closeButton.addEventListener("click", evt => {
-        evt.preventDefault();
-
-        close();
-      });
-    }
-  }
-
-  open() {
-    if (!this._element.classList.contains("popup--closed")) {
-      return;
-    }
-    this._element.classList.remove("popup--closed");
-  }
-
-  close() {
-    if (this._element.classList.contains("popup--closed")) {
-      return;
-    }
-    this._element.classList.add("popup--closed");
-  }
-
-  shake() {
-    this._element.style.animationName = "shake";
-    this._element.style.animationDuration = "0.6s";
-
-    setTimeout(() => {
-      this._element.style.animationName = "";
-      this._element.style.animationDuration = "";
-    }, 600);
-  }
-}
-
-const noopStorage = {
-  setItem() {},
-  getItem() {
-    return "";
-  }
-};
-
-if (!localStorage) {
-  localStorage = noopStorage;
-}
-
-const nameField = new Field(document.getElementById("feedback__name"));
-nameField.setRequired();
-nameField.value = localStorage.getItem("name");
-
-const emailField = new Field(document.getElementById("feedback__email"));
-emailField.setRequired();
-emailField.value = localStorage.getItem("email");
-
-const textField = new Field(document.getElementById("feedback__text"));
-textField.setRequired();
-
-const feedbackForm = new Form(document.querySelector(".feedback__form"));
-feedbackForm.addFields(nameField, emailField, textField);
-
-const feedbackModalOpenButton = document.querySelector(".contacts__link");
-const feedbackModal = new Modal(document.querySelector(".feedback.popup"));
-
-feedbackModalOpenButton.addEventListener("click", evt => {
-  evt.preventDefault();
-
-  feedbackModal.open();
-
-  feedbackForm.fields.find(f => !f.value).focus();
-});
-
-feedbackForm.addSubmitHandler(evt => {
-  if (!feedbackForm.valid()) {
-    evt.preventDefault();
-
-    return feedbackModal.shake();
-  }
-  localStorage.setItem("name", nameField.value);
-  localStorage.setItem("email", emailField.value);
-});
-
-const mapModal = new Modal(document.querySelector(".map.popup"));
-const mapModalOpenButton = document.querySelector(".contacts__map-link");
-
-mapModalOpenButton.addEventListener("click", evt => {
-  evt.preventDefault();
-
-  mapModal.open();
-});
+"use strict";function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor)){throw new TypeError("Cannot call a class as a function")}}function _defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||false;descriptor.configurable=true;if("value"in descriptor)descriptor.writable=true;Object.defineProperty(target,descriptor.key,descriptor)}}function _createClass(Constructor,protoProps,staticProps){if(protoProps)_defineProperties(Constructor.prototype,protoProps);if(staticProps)_defineProperties(Constructor,staticProps);return Constructor}var Field=/*#__PURE__*/function(){function Field(element){var _this=this;_classCallCheck(this,Field);if(!(this instanceof Field?this.constructor:void 0)){throw new Error("Field must be called with a new operator")}if(!element){throw new Error("new Field: element is empty")}this._element=element;this._validators=[function(element){return element.validity.valid}];this._element.addEventListener("keyup",function(){_this.valid()});this._element.addEventListener("change",function(){_this.valid()})}_createClass(Field,[{key:"setRequired",value:function setRequired(){var _this2=this;this._validators.unshift(function(){return _this2._element.value})}},{key:"valid",value:function valid(){var _this3=this;var isValid=this._validators.every(function(v){return v(_this3._element)});this._element.classList.remove("input__field--invalid");if(!isValid){this._element.classList.add("input__field--invalid")}return isValid}},{key:"focus",value:function focus(){return this._element.focus()}},{key:"addChangeHandler",value:function addChangeHandler(handler){this._element.addEventListener("change",handler)}},{key:"value",get:function get(){return this._element.value},set:function set(value){this._element.value=value}}]);return Field}();var Form=/*#__PURE__*/function(){function Form(element){_classCallCheck(this,Form);if(!(this instanceof Form?this.constructor:void 0)){throw new Error("Form must be called with a new operator")}if(!element){throw new Error("new Form: element is empty")}if(!(element instanceof HTMLFormElement)){throw new Error("new Form: element is not a form element")}element.noValidate=true;this._element=element;this._fields=[]}_createClass(Form,[{key:"addFields",value:function addFields(){var _this$_fields;(_this$_fields=this._fields).push.apply(_this$_fields,arguments)}},{key:"valid",value:function valid(){var valid=true;this._fields.forEach(function(f){return valid&=f.valid()});return valid}},{key:"addSubmitHandler",value:function addSubmitHandler(handler){this._element.addEventListener("submit",handler)}},{key:"fields",get:function get(){return this._fields}}]);return Form}();var Modal=/*#__PURE__*/function(){function Modal(element){_classCallCheck(this,Modal);if(!(this instanceof Modal?this.constructor:void 0)){throw new Error("Modal must be called with a new operator")}this._element=element;this._initClose()}_createClass(Modal,[{key:"_initClose",value:function _initClose(){var close=this.close.bind(this);window.addEventListener("keydown",function(evt){if(evt.keyCode!==27){return}evt.preventDefault();close()});var closeButton=this._element.querySelector(".popup__close");if(closeButton){closeButton.addEventListener("click",function(evt){evt.preventDefault();close()})}}},{key:"open",value:function open(){if(!this._element.classList.contains("popup--closed")){return}this._element.classList.remove("popup--closed")}},{key:"close",value:function close(){if(this._element.classList.contains("popup--closed")){return}this._element.classList.add("popup--closed")}},{key:"shake",value:function shake(){var _this4=this;this._element.style.animationName="shake";this._element.style.animationDuration="0.6s";setTimeout(function(){_this4._element.style.animationName="";_this4._element.style.animationDuration=""},600)}}]);return Modal}();var noopStorage={setItem:function setItem(){},getItem:function getItem(){return""}};if(!localStorage){localStorage=noopStorage}var nameField=new Field(document.getElementById("feedback__name"));nameField.setRequired();nameField.value=localStorage.getItem("name");var emailField=new Field(document.getElementById("feedback__email"));emailField.setRequired();emailField.value=localStorage.getItem("email");var textField=new Field(document.getElementById("feedback__text"));textField.setRequired();var feedbackForm=new Form(document.querySelector(".feedback__form"));feedbackForm.addFields(nameField,emailField,textField);var feedbackModalOpenButton=document.querySelector(".contacts__link");var feedbackModal=new Modal(document.querySelector(".feedback.popup"));feedbackModalOpenButton.addEventListener("click",function(evt){evt.preventDefault();feedbackModal.open();feedbackForm.fields.find(function(f){return!f.value}).focus()});feedbackForm.addSubmitHandler(function(evt){if(!feedbackForm.valid()){evt.preventDefault();return feedbackModal.shake()}localStorage.setItem("name",nameField.value);localStorage.setItem("email",emailField.value)});var mapModal=new Modal(document.querySelector(".map.popup"));var mapModalOpenButton=document.querySelector(".contacts__map-link");mapModalOpenButton.addEventListener("click",function(evt){evt.preventDefault();mapModal.open()});
