+++ postcss.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: [
+    require("cssnano")({
+      preset: "default"
+    })
+  ]
+};
